+++ pkg/apis/configuration/v1beta1/groupversion_info.go
@@ -15,8 +15,8 @@ limitations under the License.
 */
 
 // Package v1beta1 contains API Schema definitions for the configuration.konghq.com v1beta1 API group
-//+kubebuilder:object:generate=true
-//+groupName=configuration.konghq.com
+// +kubebuilder:object:generate=true
+// +groupName=configuration.konghq.com
 package v1beta1
 
 import (